sync main to beta

+++ static/js/TasksUpdateModal.js
@@ -2,7 +2,7 @@ const TasksUpdateModal = {
     components: {
         'input-stepper': InputStepper,
     },
-    props: ['runtimes', 'selected-task'],
+    props: ['runtimes', 'selected-task', 'integrations', 's3Integrations',],
     data() {
         return this.initial_state()
     },
@@ -17,6 +17,9 @@ const TasksUpdateModal = {
             this.runtime = e.target.value;
         })
         $('#task_modal_parallel').closest('div.custom-input').hide();
+        $('#selector_integration_update_task').on('change', (e) => {
+            this.updateIntegration(e.target.value);
+        });
     },
     computed: {
         test_parameters() {
@@ -28,22 +31,30 @@ const TasksUpdateModal = {
     },
     methods: {
         setTaskData(taskData) {
+            this.selectedIntegration = `${taskData.zippath.integration_id}#${taskData.zippath.is_local}`
             this.$nextTick(() => {
                 $('#selectUpdatedRuntime').val(taskData.runtime);
                 $('#selectUpdatedRuntime').selectpicker('refresh');
+                $('#selector_integration_update_task').val(this.selectedIntegration);
+                $('#selector_integration_update_task').selectpicker('refresh');
             })
+            this.integration_id = taskData.zippath.integration_id;
+            this.is_local = taskData.zippath.is_local;
             this.runtime = taskData.runtime;
             this.task_name = taskData.task_name;
             this.task_handler = taskData.task_handler;
-            this.previewFile = taskData.zippath;
+            this.task_handler = taskData.task_handler;
+            this.previewFile = `${taskData.zippath.bucket_name}/${taskData.zippath.file_name}`;
+            this.old_file_name = taskData.zippath.file_name;
             const envVars = JSON.parse(taskData.env_vars);
+            this.monitoring_settings = envVars.monitoring_settings  || this.initial_state().monitoring_settings;
             if (envVars.task_parameters) {
                 this.test_parameters.set(envVars.task_parameters);
             }
         },
         async fetchTaskInfo() {
             const api_url = this.$root.build_api_url('tasks', 'tasks')
-            const res = await fetch (`${api_url}/${getSelectedProjectId()}/${this.selectedTask.task_id}`,{
+            const res = await fetch(`${api_url}/${getSelectedProjectId()}/${this.selectedTask.task_id}`, {
                 method: 'GET',
             })
             return res.json();
@@ -57,6 +68,14 @@ const TasksUpdateModal = {
                 previewFile: '',
                 file: null,
                 isSubmitted: false,
+                monitoring_settings: {
+                    integration: null,
+                    failed_tasks: 5,
+                    recipients: [],
+                },
+                integration_id: null,
+                is_local: false,
+                selectedIntegration: null
             }
         },
         get_data() {
@@ -65,24 +84,29 @@ const TasksUpdateModal = {
                 "task_handler": this.task_handler,
                 "runtime": this.runtime,
                 "task_package": this.previewFile,
-                "task_parameters": this.test_parameters.get()
+                "task_parameters": this.test_parameters.get(),
+                "monitoring_settings": this.monitoring_settings,
+                "s3_settings": {
+                    "integration_id": this.integration_id, 
+                    "is_local": this.is_local
+                },
             }
         },
         uploadFile(e) {
-            const file =  e.target.files[0];
+            const file = e.target.files[0];
             this.previewFile = file.name
             this.file = file
             return file
         },
         onDrop(e) {
-            const file =  e.dataTransfer.files[0];
+            const file = e.dataTransfer.files[0];
             this.previewFile = file.name
             this.file = file
             return file
         },
-        async updateTaskAPI(data){
+        async updateTaskAPI(data) {
             const api_url = this.$root.build_api_url('tasks', 'tasks')
-            const resp = await fetch(`${api_url}/${getSelectedProjectId()}/${this.selectedTask.task_id}`,{
+            const resp = await fetch(`${api_url}/${getSelectedProjectId()}/${this.selectedTask.task_id}`, {
                 method: 'PUT',
                 body: data,
             })
@@ -102,21 +126,36 @@ const TasksUpdateModal = {
                 this.isLoading = true;
                 let data = new FormData();
                 if (this.file) data.append('file', this.file);
-                const prepareData = this.file ? this.get_data() : { ...this.get_data(), "task_package": ""}
+                const prepareData = this.file ? this.get_data() : {...this.get_data(), 
+                    "task_package": this.old_file_name, "validate_package": false}
                 data.append('data', JSON.stringify(prepareData));
-                this.updateTaskAPI(data).then( response => {
+                this.updateTaskAPI(data).then(response => {
+                    if (response[0]?.type === "assertion_error") {
+                        throw new Error(response[0].msg)
+                    }
                     showNotify('SUCCESS', 'Task Updated.');
                     this.$emit('update-tasks-list', response.task_id);
                     $('#UpdateTaskModal').modal('hide');
                     form.classList.remove('was-validated');
                     this.removeFile();
                 })
-                .catch(err => {
-                    showNotify('ERROR', err);
-                }).finally(() => {
+                    .catch(err => {
+                        console.error(err)
+                        showNotify('ERROR', err);
+                    }).finally(() => {
                     this.isLoading = false;
                 })
             }
+        },
+        get_integration_value(integration) {
+            return `${integration?.id}#${!!(integration?.project_id)}`
+        },
+        getIntegrationTitle(integration) {
+            return integration.is_default ? `${integration.config?.name} - default` : integration.config?.name
+        },
+        updateIntegration(integration_value) {
+            this.integration_id = parseInt(integration_value?.split('#')[0])
+            this.is_local = integration_value?.split('#')[1] === 'true'
         },
     },
     template: `
@@ -166,11 +205,26 @@ const TasksUpdateModal = {
                                         {{ previewFile }}
                                         <i class="icon__16x16 icon-close__16" @click="removeFile"></i>
                                     </span>
+                                    <p class="font-h5 font-bold">Storage</p>
+                                    <p class="font-h6 font-weight-400">Choose your S3 storage to save your task package</p>
+                                    <div class="w-100-imp">
+                                        <select id='selector_integration_update_task' class="selectpicker bootstrap-select__b mb-3 mt-2" data-style="btn"
+                                            >
+                                            <option
+                                                v-for="integration in s3Integrations"
+                                                :value="get_integration_value(integration)"
+                                                :title="getIntegrationTitle(integration)"
+                                                :key="integration"
+                                            >
+                                                {{ getIntegrationTitle(integration) }}
+                                            </option>
+                                        </select>
+                                    </div> 
                                 </div>
                                 <div class="form-group">
                                     <p class="font-h5 font-bold">Runtime</p>
                                     <p class="font-h6 font-weight-400">Choose the language to use to write your function</p>
-                                    <div class=" w-100-imp">
+                                    <div class="w-100-imp">
                                         <select class="selectpicker bootstrap-select__need-validation mb-3 mt-2" 
                                             id="selectUpdatedRuntime"
                                             data-style="btn"
@@ -195,6 +249,10 @@ const TasksUpdateModal = {
                                 </div>
                             </form>
                             <slot></slot>
+                            <tasks-monitoring v-if="integrations.length > 0"
+                                :integrations="integrations"
+                                :monitoring_settings="monitoring_settings"
+                            ></tasks-monitoring>
                         </div>
                     </div>
                 </div>
